add performance benchmark
